steganography/requirements.txt
pyaes=1.6.1
Pillow=9.4.0